Improve robustness, flexibility and documentation
This commit introduces end-user options such as '--overwrite' to allow
the overwriting of data during a restoration procedure.

It also introduces a bit of robustness, by catching exceptions and, for
example, backing up data if '--overwrite' is used, to avoid destructive
outcomes.

The README has been updated to reflect the current functionality and
invocation parameters.